It is done, sort of.  compiles as C++ atleast and runs the game without sound and net code.   Not really const safe yet so going to use a string ref class and buffers.  Still debating to use stl or not.

+++ quake_framebuffer/quake_framebuffer/zone.h
@@ -85,20 +85,20 @@ Zone block
 #ifndef _QUAKE_ZONE_H_
 #define _QUAKE_ZONE_H_
 
-void Memory_Init (void *buf, int size);
+void Memory_Init (void *buf, size_t size);
 
 void Z_Free (void *ptr);
-void *Z_Malloc (int size);			// returns 0 filled memory
-void *Z_TagMalloc (int size, int tag);
+void *Z_Malloc (size_t size);			// returns 0 filled memory
+void *Z_TagMalloc (size_t size, int tag);
 
 void Z_DumpHeap (void);
 void Z_CheckHeap (void);
 int Z_FreeMemory (void);
 
-void *Hunk_Alloc (int size);		// returns 0 filled memory
+void *Hunk_Alloc (size_t size);		// returns 0 filled memory
 void *Hunk_AllocName (int size, char *name);
 
-void *Hunk_HighAllocName (int size, char *name);
+void *Hunk_HighAllocName (size_t size, char *name);
 
 int	Hunk_LowMark (void);
 void Hunk_FreeToLowMark (int mark);
@@ -106,7 +106,7 @@ void Hunk_FreeToLowMark (int mark);
 int	Hunk_HighMark (void);
 void Hunk_FreeToHighMark (int mark);
 
-void *Hunk_TempAlloc (int size);
+void *Hunk_TempAlloc (size_t size);
 
 void Hunk_Check (void);
